Programming, Ops and Database Exercise/Python API/requirements.txt
flask
mysql-connector
dotenv